keep track of object and action params seperately
